pyproject.toml
[project]
name = "social-graph-py"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "neo4j>=6.0.2",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.20.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
